[EI-227] [web] feat: Add SplitScreenToggleGroup component
